Implemented add/remove product

+++ src/main/java/com/example/persistence/model/Order.java
@@ -4,6 +4,8 @@ import io.quarkus.mongodb.panache.PanacheMongoEntity;
 import io.quarkus.mongodb.panache.common.MongoEntity;
 import org.bson.codecs.pojo.annotations.BsonProperty;
 import org.bson.types.ObjectId;
+
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -20,7 +22,7 @@ public class Order extends PanacheMongoEntity {
     @BsonProperty("status")
     public String status;
     @BsonProperty("price")
-    public Double price;
+    public BigDecimal price;
     @BsonProperty("order_date")
     public Date orderDate;
     @BsonProperty("pickup_date")
@@ -38,7 +40,7 @@ public class Order extends PanacheMongoEntity {
     public Order() {
 
     }
-    public Order(ObjectId id, List<ProductItem> products, String email, String phone, String status, Double price, Date orderDate, Date pickupDate) {
+    public Order(ObjectId id, List<ProductItem> products, String email, String phone, String status, BigDecimal price, Date orderDate, Date pickupDate) {
         this.id = id;
         this.products = products;
         this.email = email;
@@ -89,11 +91,11 @@ public class Order extends PanacheMongoEntity {
         this.status = status;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
